requirements.txt
requests
pandas
mysql-connector-python
argparse
logging